CONTCNTR-5084 and CONTCNTR-5082
