+++ pkg/controller/postManager.go
@@ -24,7 +24,6 @@ import (
 	"fmt"
 	"io"
 	"net/http"
-	"strconv"
 	"strings"
 	"time"
 
@@ -47,7 +46,6 @@ func NewPostManager(params AgentParams, gtmPostMgr bool) *PostManager {
 		retryTenantDeclMap:              make(map[string]*tenantParams),
 		tenantPriorityMap:               make(map[string]int),
 		postChan:                        make(chan ResourceConfigRequest, 1),
-		retryChan:                       make(chan struct{}, 1),
 	}
 	if !gtmPostMgr {
 		pm.PostParams = params.PostParams
@@ -101,33 +99,13 @@ func (postMgr *PostManager) setupBIGIPRESTClient() {
 	}
 }
 
-func (postMgr *PostManager) getAS3APIURL(tenants []string) string {
-	apiURL := postMgr.BIGIPURL + "/mgmt/shared/appsvcs/declare/" + strings.Join(tenants, ",")
-	return apiURL
-}
-
-func (postMgr *PostManager) getAS3TaskIdURL(taskId string) string {
-	apiURL := postMgr.BIGIPURL + "/mgmt/shared/appsvcs/task/" + taskId
-	return apiURL
-}
-
-// publishConfig posts incoming configuration to BIG-IP
-func (postMgr *PostManager) publishConfig(cfg agentConfig) {
-	log.Debugf("[AS3]%v PostManager Accepted the configuration", postMgr.postManagerPrefix)
-	// postConfig updates the tenantResponseMap with response codes
-	postMgr.postConfig(&cfg)
-}
-
-func (postMgr *PostManager) postConfig(cfg *agentConfig) {
+func (postMgr *PostManager) postConfig(cfg *agentConfig) (*http.Response, map[string]interface{}) {
 	// log as3 request if it's set
-	if postMgr.LogAS3Request {
-		postMgr.logAS3Request(cfg.data)
-	}
 	httpReqBody := bytes.NewBuffer([]byte(cfg.data))
 	req, err := http.NewRequest("POST", cfg.as3APIURL, httpReqBody)
 	if err != nil {
 		log.Errorf("%v[AS3]%v Creating new HTTP request error: %v ", getRequestPrefix(cfg.id), postMgr.postManagerPrefix, err)
-		return
+		return nil, nil
 	}
 	log.Debugf("[AS3]%v posting request to %v", postMgr.postManagerPrefix, cfg.as3APIURL)
 	log.Infof("%v[AS3]%v posting request to %v for %v tenants", getRequestPrefix(cfg.id), postMgr.postManagerPrefix, postMgr.BIGIPURL, getTenantsFromUri(cfg.as3APIURL))
@@ -135,36 +113,13 @@ func (postMgr *PostManager) postConfig(cfg *agentConfig) {
 
 	httpResp, responseMap := postMgr.httpPOST(req)
 	if httpResp == nil || responseMap == nil {
-		return
+		return nil, nil
 	}
 
 	if postMgr.firstPost {
 		postMgr.firstPost = false
 	}
-
-	switch httpResp.StatusCode {
-	case http.StatusOK:
-		log.Infof("%v[AS3]%v post resulted in SUCCESS", getRequestPrefix(cfg.id), postMgr.postManagerPrefix)
-		postMgr.handleResponseStatusOK(responseMap)
-	case http.StatusCreated, http.StatusAccepted:
-		log.Infof("%v[AS3]%v post resulted in ACCEPTED", getRequestPrefix(cfg.id), postMgr.postManagerPrefix)
-		postMgr.handleResponseAccepted(responseMap)
-	case http.StatusMultiStatus:
-		log.Infof("%v[AS3]%v post resulted in MULTI-STATUS", getRequestPrefix(cfg.id), postMgr.postManagerPrefix)
-		postMgr.handleMultiStatus(responseMap, cfg.id)
-	case http.StatusServiceUnavailable:
-		log.Infof("%v[AS3]%v post resulted in RETRY", getRequestPrefix(cfg.id), postMgr.postManagerPrefix)
-		postMgr.handleResponseStatusServiceUnavailable(responseMap, cfg.id)
-	case http.StatusNotFound:
-		log.Infof("%v[AS3]%v post resulted in FAILURE", getRequestPrefix(cfg.id), postMgr.postManagerPrefix)
-		postMgr.handleResponseStatusNotFound(responseMap, cfg.id)
-	case http.StatusUnauthorized:
-		log.Infof("%v[AS3]%v post resulted in UNAUTHORIZED FAILURE", getRequestPrefix(cfg.id), postMgr.postManagerPrefix)
-		postMgr.handleResponseStatusUnAuthorized(responseMap, cfg.id)
-	default:
-		log.Infof("%v[AS3]%v post resulted in FAILURE", getRequestPrefix(cfg.id), postMgr.postManagerPrefix)
-		postMgr.handleResponseOthers(responseMap, cfg.id)
-	}
+	return httpResp, responseMap
 }
 
 func updateTenantDeletion(tenant string, declaration map[string]interface{}) bool {
@@ -196,7 +151,7 @@ func (postMgr *PostManager) httpPOST(request *http.Request) (*http.Response, map
 		if httpResp.StatusCode == http.StatusUnauthorized {
 			log.Errorf("[AS3]%v Unauthorized access to BIG-IP, please check the credentials, message: %v", postMgr.postManagerPrefix, string(body))
 		}
-		if postMgr.LogAS3Response {
+		if postMgr.LogResponse {
 			log.Errorf("[AS3]%v Raw response from Big-IP: %v", postMgr.postManagerPrefix, string(body))
 		}
 		return nil, nil
@@ -215,7 +170,7 @@ func (postMgr *PostManager) updateTenantResponseCode(code int, id string, tenant
 	}
 }
 
-func (postMgr *PostManager) handleResponseStatusOK(responseMap map[string]interface{}) {
+func (postMgr *PostManager) handleResponseStatusOK(responseMap map[string]interface{}) bool {
 	var unknownResponse bool
 	// traverse all response results
 	results, ok1 := (responseMap["results"]).([]interface{})
@@ -238,73 +193,10 @@ func (postMgr *PostManager) handleResponseStatusOK(responseMap map[string]interf
 	} else {
 		unknownResponse = true
 	}
-	if postMgr.LogAS3Response || unknownResponse {
-		postMgr.logAS3Response(responseMap)
-	}
+	return unknownResponse
 }
 
-func (postMgr *PostManager) getTenantConfigStatus(id string) {
-	req, err := http.NewRequest("GET", postMgr.getAS3TaskIdURL(id), nil)
-	if err != nil {
-		log.Errorf("[AS3]%v Creating new HTTP request error: %v ", postMgr.postManagerPrefix, err)
-		return
-	}
-	log.Debugf("[AS3]%v posting request with taskId to %v", postMgr.postManagerPrefix, postMgr.getAS3TaskIdURL(id))
-	req.SetBasicAuth(postMgr.BIGIPUsername, postMgr.BIGIPPassword)
-
-	httpResp, responseMap := postMgr.httpPOST(req)
-	if httpResp == nil || responseMap == nil {
-		return
-	}
-
-	if postMgr.LogAS3Response {
-		postMgr.logAS3Response(responseMap)
-	}
-	var unknownResponse bool
-	if httpResp.StatusCode == http.StatusOK {
-		results, ok1 := (responseMap["results"]).([]interface{})
-		declaration, ok2 := (responseMap["declaration"]).(interface{}).(map[string]interface{})
-		if ok1 && ok2 {
-			for _, value := range results {
-				if v, ok := value.(map[string]interface{}); ok {
-					code, ok1 := v["code"].(float64)
-					tenant, ok2 := v["tenant"].(string)
-					msg, ok3 := v["message"]
-					if ok1 && ok2 && ok3 {
-						if message, ok := msg.(string); ok && message == "in progress" {
-							return
-						}
-						// reset task id, so that any unknownResponse failed will go to post call in the next retry
-						postMgr.updateTenantResponseCode(int(code), "", tenant, updateTenantDeletion(tenant, declaration), "")
-						if _, ok := v["response"]; ok {
-							log.Debugf("[AS3]%v Response from BIG-IP: code: %v --- tenant:%v --- message: %v %v", postMgr.postManagerPrefix, v["code"], v["tenant"], v["message"], v["response"])
-						} else {
-							log.Debugf("[AS3]%v Response from BIG-IP: code: %v --- tenant:%v --- message: %v", postMgr.postManagerPrefix, v["code"], v["tenant"], v["message"])
-						}
-						intId, err := strconv.Atoi(id)
-						if err == nil {
-							log.Infof("%v[AS3]%v post resulted in SUCCESS", getRequestPrefix(intId), postMgr.postManagerPrefix)
-						}
-					} else {
-						unknownResponse = true
-					}
-				} else {
-					unknownResponse = true
-				}
-			}
-		} else {
-			unknownResponse = true
-		}
-	} else if httpResp.StatusCode != http.StatusServiceUnavailable {
-		// reset task id, so that any failed tenants will go to post call in the next retry
-		postMgr.updateTenantResponseCode(httpResp.StatusCode, "", "", false, "")
-	}
-	if !postMgr.LogAS3Response && unknownResponse {
-		postMgr.logAS3Response(responseMap)
-	}
-}
-
-func (postMgr *PostManager) handleMultiStatus(responseMap map[string]interface{}, id int) {
+func (postMgr *PostManager) handleMultiStatus(responseMap map[string]interface{}, id int) bool {
 	var unknownResponse bool
 	results, ok1 := (responseMap["results"]).([]interface{})
 	declaration, ok2 := (responseMap["declaration"]).(interface{}).(map[string]interface{})
@@ -331,34 +223,34 @@ func (postMgr *PostManager) handleMultiStatus(responseMap map[string]interface{}
 	} else {
 		unknownResponse = true
 	}
-	if postMgr.LogAS3Response || unknownResponse {
-		postMgr.logAS3Response(responseMap)
-	}
+	return unknownResponse
 }
 
-func (postMgr *PostManager) handleResponseAccepted(responseMap map[string]interface{}) {
+func (postMgr *PostManager) handleResponseAccepted(responseMap map[string]interface{}) bool {
 	// traverse all response results
+	var unknownResponse bool
 	if respId, ok := (responseMap["id"]).(string); ok {
 		postMgr.updateTenantResponseCode(http.StatusAccepted, respId, "", false, "")
 		log.Debugf("[AS3]%v Response from BIG-IP: code 201 id %v, waiting %v seconds to poll response", postMgr.postManagerPrefix, respId, timeoutMedium)
-	} else {
-		postMgr.logAS3Response(responseMap)
+		unknownResponse = true
 	}
+	return unknownResponse
 }
 
-func (postMgr *PostManager) handleResponseStatusServiceUnavailable(responseMap map[string]interface{}, id int) {
+func (postMgr *PostManager) handleResponseStatusServiceUnavailable(responseMap map[string]interface{}, id int) bool {
 	var message string
+	var unknownResponse bool
 	if err, ok := (responseMap["error"]).(map[string]interface{}); ok {
 		log.Errorf("%v[AS3]%v Big-IP Responded with error code: %v", getRequestPrefix(id), postMgr.postManagerPrefix, err["code"])
 		message = fmt.Sprintf("Big-IP Responded with error code: %v -- verify the logs for detailed error", err["code"])
-	} else {
-		postMgr.logAS3Response(responseMap)
+		unknownResponse = true
 	}
 	log.Debugf("[AS3]%v Response from BIG-IP: BIG-IP is busy, waiting %v seconds and re-posting the declaration", postMgr.postManagerPrefix, timeoutMedium)
 	postMgr.updateTenantResponseCode(http.StatusServiceUnavailable, "", "", false, message)
+	return unknownResponse
 }
 
-func (postMgr *PostManager) handleResponseStatusNotFound(responseMap map[string]interface{}, id int) {
+func (postMgr *PostManager) handleResponseStatusNotFound(responseMap map[string]interface{}, id int) bool {
 	var unknownResponse bool
 	var message string
 	if err, ok := (responseMap["error"]).(map[string]interface{}); ok {
@@ -368,13 +260,11 @@ func (postMgr *PostManager) handleResponseStatusNotFound(responseMap map[string]
 		unknownResponse = true
 		message = "Big-IP Responded with error -- verify the logs for detailed error"
 	}
-	if postMgr.LogAS3Response || unknownResponse {
-		postMgr.logAS3Response(responseMap)
-	}
 	postMgr.updateTenantResponseCode(http.StatusNotFound, "", "", false, message)
+	return unknownResponse
 }
 
-func (postMgr *PostManager) handleResponseStatusUnAuthorized(responseMap map[string]interface{}, id int) {
+func (postMgr *PostManager) handleResponseStatusUnAuthorized(responseMap map[string]interface{}, id int) bool {
 	var unknownResponse bool
 	var message string
 	if _, ok := responseMap["code"].(float64); ok {
@@ -391,13 +281,11 @@ func (postMgr *PostManager) handleResponseStatusUnAuthorized(responseMap map[str
 		message = "Big-IP Responded with error -- verify the logs for detailed error"
 	}
 
-	if postMgr.LogAS3Response || unknownResponse {
-		postMgr.logAS3Response(responseMap)
-	}
 	postMgr.updateTenantResponseCode(http.StatusUnauthorized, "", "", false, message)
+	return unknownResponse
 }
 
-func (postMgr *PostManager) handleResponseOthers(responseMap map[string]interface{}, id int) {
+func (postMgr *PostManager) handleResponseOthers(responseMap map[string]interface{}, id int) bool {
 	var unknownResponse bool
 	if results, ok := (responseMap["results"]).([]interface{}); ok {
 		for _, value := range results {
@@ -427,175 +315,7 @@ func (postMgr *PostManager) handleResponseOthers(responseMap map[string]interfac
 			postMgr.updateTenantResponseCode(int(code), "", "", false, fmt.Sprintf("Big-IP Responded with error code: %v -- verify the logs for detailed error", code))
 		}
 	}
-	if postMgr.LogAS3Response || unknownResponse {
-		postMgr.logAS3Response(responseMap)
-	}
-}
-
-func (postMgr *PostManager) GetBigipAS3Version() (string, string, string, error) {
-	url := postMgr.getAS3VersionURL()
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		log.Errorf("[AS3]%v Creating new HTTP request error: %v ", postMgr.postManagerPrefix, err)
-		return "", "", "", err
-	}
-
-	log.Debugf("[AS3]%v posting GET BIGIP AS3 Version request on %v", postMgr.postManagerPrefix, url)
-	req.SetBasicAuth(postMgr.BIGIPUsername, postMgr.BIGIPPassword)
-
-	httpResp, responseMap := postMgr.httpReq(req)
-	if httpResp == nil || responseMap == nil {
-		return "", "", "", fmt.Errorf("Internal Error")
-	}
-	var unknownResponse bool
-	switch httpResp.StatusCode {
-	case http.StatusOK:
-		if responseMap["version"] != nil {
-			as3VersionStr, ok1 := responseMap["version"].(string)
-			as3versionreleaseStr, ok2 := responseMap["release"].(string)
-			as3SchemaVersion, ok3 := responseMap["schemaCurrent"].(string)
-			if ok1 && ok2 && ok3 {
-				return as3VersionStr, as3versionreleaseStr, as3SchemaVersion, nil
-			} else {
-				unknownResponse = true
-			}
-		} else {
-			unknownResponse = true
-		}
-	case http.StatusNotFound:
-		if code, ok := responseMap["code"].(float64); ok {
-			if code == http.StatusNotFound {
-				return "", "", "", fmt.Errorf("AS3 RPM is not installed on BIGIP,"+
-					" Error response from BIGIP with status code %v", httpResp.StatusCode)
-			}
-		} else {
-			unknownResponse = true
-		}
-		// In case of 503 status code : CIS will exit and auto restart of the
-		// controller might fetch the BIGIP version once BIGIP is available.
-	case http.StatusUnauthorized:
-		if code, ok := responseMap["code"].(float64); ok {
-			if int(code) == http.StatusUnauthorized {
-				if _, ok := responseMap["message"].(string); ok {
-					return "", "", "", fmt.Errorf("authentication failed,"+
-						" Error response from BIGIP with status code %v Message: %v", httpResp.StatusCode, responseMap["message"])
-				} else {
-					return "", "", "", fmt.Errorf("authentication failed,"+
-						" Error response from BIGIP with status code %v", httpResp.StatusCode)
-				}
-			}
-		} else {
-			unknownResponse = true
-		}
-	}
-	if unknownResponse {
-		postMgr.logAS3Response(responseMap)
-	}
-	return "", "", "", fmt.Errorf("Error response from BIGIP with status code %v", httpResp.StatusCode)
-}
-
-// GetBigipRegKey ...
-func (postMgr *PostManager) GetBigipRegKey() (string, error) {
-	url := postMgr.getBigipRegKeyURL()
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		log.Errorf("[AS3]%v Creating new HTTP request error: %v ", postMgr.postManagerPrefix, err)
-		return "", err
-	}
-
-	log.Debugf("[AS3]%v Posting GET BIGIP Reg Key request on %v", postMgr.postManagerPrefix, url)
-	req.SetBasicAuth(postMgr.BIGIPUsername, postMgr.BIGIPPassword)
-
-	httpResp, responseMap := postMgr.httpReq(req)
-	if httpResp == nil || responseMap == nil {
-		return "", fmt.Errorf("Internal Error")
-	}
-	var unknownResponse bool
-	switch httpResp.StatusCode {
-	case http.StatusOK:
-		if regKey, ok := responseMap["registrationKey"]; ok {
-			if registrationKey, ok := regKey.(string); ok {
-				return registrationKey, nil
-			} else {
-				unknownResponse = true
-			}
-		} else {
-			unknownResponse = true
-		}
-	case http.StatusNotFound:
-		if code, ok := responseMap["code"].(float64); ok {
-			if int(code) == http.StatusNotFound {
-				return "", fmt.Errorf("AS3 RPM is not installed on BIGIP,"+
-					" Error response from BIGIP with status code %v", httpResp.StatusCode)
-			}
-		} else {
-			unknownResponse = true
-		}
-	case http.StatusUnauthorized:
-		if code, ok := responseMap["code"].(float64); ok {
-			if int(code) == http.StatusUnauthorized {
-				if _, ok := responseMap["message"].(string); ok {
-					return "", fmt.Errorf("authentication failed,"+
-						" Error response from BIGIP with status code %v Message: %v", httpResp.StatusCode, responseMap["message"])
-				} else {
-					return "", fmt.Errorf("authentication failed,"+
-						" Error response from BIGIP with status code %v", httpResp.StatusCode)
-				}
-			}
-		} else {
-			unknownResponse = true
-		}
-	}
-	if unknownResponse {
-		postMgr.logAS3Response(responseMap)
-	}
-	return "", fmt.Errorf("Error response from BIGIP with status code %v", httpResp.StatusCode)
-}
-
-func (postMgr *PostManager) GetAS3DeclarationFromBigIP() (map[string]interface{}, error) {
-	url := postMgr.getAS3APIURL([]string{})
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		log.Errorf("[AS3]%v Creating new HTTP request error: %v ", postMgr.postManagerPrefix, err)
-		return nil, err
-	}
-
-	log.Debugf("[AS3]%v posting GET BIGIP AS3 declaration request on %v", postMgr.postManagerPrefix, url)
-	req.SetBasicAuth(postMgr.BIGIPUsername, postMgr.BIGIPPassword)
-
-	httpResp, responseMap := postMgr.httpReq(req)
-	if httpResp == nil || responseMap == nil {
-		return nil, fmt.Errorf("Internal Error")
-	}
-
-	switch httpResp.StatusCode {
-	case http.StatusOK:
-		return responseMap, err
-	case http.StatusNotFound:
-		if code, ok := responseMap["code"].(float64); ok {
-			if int(code) == http.StatusNotFound {
-				return nil, fmt.Errorf("AS3 RPM is not installed on BIGIP,"+
-					" Error response from BIGIP with status code %v", httpResp.StatusCode)
-			}
-		} else {
-			postMgr.logAS3Response(responseMap)
-		}
-	case http.StatusUnauthorized:
-		if code, ok := responseMap["code"].(float64); ok {
-			if int(code) == http.StatusUnauthorized {
-				if _, ok := responseMap["message"].(string); ok {
-					return nil, fmt.Errorf("authentication failed,"+
-						" Error response from BIGIP with status code %v Message: %v", httpResp.StatusCode, responseMap["message"])
-				} else {
-					return nil, fmt.Errorf("authentication failed,"+
-						" Error response from BIGIP with status code %v", httpResp.StatusCode)
-				}
-			}
-		} else {
-			postMgr.logAS3Response(responseMap)
-		}
-	}
-	return nil, fmt.Errorf("Error response from BIGIP with status code %v", httpResp.StatusCode)
+	return unknownResponse
 }
 
 func (postMgr *PostManager) httpReq(request *http.Request) (*http.Response, map[string]interface{}) {
@@ -618,7 +338,7 @@ func (postMgr *PostManager) httpReq(request *http.Request) (*http.Response, map[
 		if httpResp.StatusCode == http.StatusUnauthorized {
 			log.Errorf("[AS3]%v Unauthorized access to BIG-IP, please check the credentials, message: %v", postMgr.postManagerPrefix, string(body))
 		}
-		if postMgr.LogAS3Response {
+		if postMgr.LogResponse {
 			log.Errorf("[AS3]%v Raw response from Big-IP: %v", postMgr.postManagerPrefix, string(body))
 		}
 		return nil, nil
@@ -626,119 +346,16 @@ func (postMgr *PostManager) httpReq(request *http.Request) (*http.Response, map[
 	return httpResp, response
 }
 
-func (postMgr *PostManager) getAS3VersionURL() string {
-	apiURL := postMgr.BIGIPURL + "/mgmt/shared/appsvcs/info"
-	return apiURL
-}
+// GetBigipRegKey ...
 
-func (postMgr *PostManager) getBigipRegKeyURL() string {
-	apiURL := postMgr.BIGIPURL + "/mgmt/tm/shared/licensing/registration"
-	return apiURL
-}
-
-func (postMgr *PostManager) logAS3Response(responseMap map[string]interface{}) {
-	// removing the certificates/privateKey from response log
-	if declaration, ok := (responseMap["declaration"]).([]interface{}); ok {
-		for _, value := range declaration {
-			if tenantMap, ok := value.(map[string]interface{}); ok {
-				for _, value2 := range tenantMap {
-					if appMap, ok := value2.(map[string]interface{}); ok {
-						for _, obj := range appMap {
-							if crt, ok := obj.(map[string]interface{}); ok {
-								if crt["class"] == "Certificate" {
-									crt["certificate"] = ""
-									crt["privateKey"] = ""
-									crt["chainCA"] = ""
-								}
-							}
-						}
-					}
-				}
-			}
-		}
-		decl, err := json.Marshal(declaration)
-		if err != nil {
-			log.Errorf("[AS3]%v error while reading declaration from AS3 response: %v\n", postMgr.postManagerPrefix, err)
-			return
-		}
-		responseMap["declaration"] = as3Declaration(decl)
-	}
-	log.Debugf("[AS3]%v Raw response from Big-IP: %v ", postMgr.postManagerPrefix, responseMap)
-}
-
-func (postMgr *PostManager) logAS3Request(cfg string) {
-	var as3Config map[string]interface{}
-	err := json.Unmarshal([]byte(cfg), &as3Config)
-	if err != nil {
-		log.Errorf("[AS3]%v Request body unmarshal failed: %v\n", postMgr.postManagerPrefix, err)
-	}
-	adc := as3Config["declaration"].(map[string]interface{})
-	for _, value := range adc {
-		if tenantMap, ok := value.(map[string]interface{}); ok {
-			for _, value2 := range tenantMap {
-				if appMap, ok := value2.(map[string]interface{}); ok {
-					for _, obj := range appMap {
-						if crt, ok := obj.(map[string]interface{}); ok {
-							if crt["class"] == "Certificate" {
-								crt["certificate"] = ""
-								crt["privateKey"] = ""
-								crt["chainCA"] = ""
-							}
-						}
-					}
-				}
-			}
-		}
-	}
-	decl, err := json.Marshal(as3Config)
-	if err != nil {
-		log.Errorf("[AS3]%v Unified declaration error: %v\n", postMgr.postManagerPrefix, err)
-		return
-	}
-	log.Debugf("[AS3]%v Unified declaration: %v\n", postMgr.postManagerPrefix, as3Declaration(decl))
-}
+//func (postMgr *PostManager) getBigipRegKeyURL() string {
+//	apiURL := postMgr.BIGIPURL + "/mgmt/tm/shared/licensing/registration"
+//	return apiURL
+//}
 
 // Method to verify if App Services are installed or CIS as3 version is
 // compatible with BIG-IP, it will return with error if any one of the
 // requirements are not met
-func (postMgr *PostManager) IsBigIPAppServicesAvailable() error {
-	version, build, schemaVersion, err := postMgr.GetBigipAS3Version()
-	if err != nil {
-		log.Errorf("[AS3]%v %v ", postMgr.postManagerPrefix, err)
-		return err
-	}
-	am := as3VersionInfo{
-		as3Version:       version,
-		as3SchemaVersion: schemaVersion,
-		as3Release:       version + "-" + build,
-	}
-	postMgr.AS3VersionInfo = am
-	versionstr := version[:strings.LastIndex(version, ".")]
-	bigIPAS3Version, err := strconv.ParseFloat(versionstr, 64)
-	if err != nil {
-		log.Errorf("[AS3]%v Error while converting AS3 version to float", postMgr.postManagerPrefix)
-		return err
-	}
-	postMgr.bigIPAS3Version = bigIPAS3Version
-	if bigIPAS3Version >= as3SupportedVersion && bigIPAS3Version <= as3Version {
-		log.Debugf("[AS3]%v BIGIP is serving with AS3 version: %v", postMgr.postManagerPrefix, version)
-		return nil
-	}
-
-	if bigIPAS3Version > as3Version {
-		am.as3Version = defaultAS3Version
-		am.as3SchemaVersion = fmt.Sprintf("%.2f.0", as3Version)
-		as3Build := defaultAS3Build
-		am.as3Release = am.as3Version + "-" + as3Build
-		log.Debugf("[AS3]%v BIGIP is serving with AS3 version: %v", postMgr.postManagerPrefix, bigIPAS3Version)
-		postMgr.AS3VersionInfo = am
-		return nil
-	}
-
-	return fmt.Errorf("CIS versions >= 2.0 are compatible with AS3 versions >=%v. "+
-		"Upgrade AS3 version in BIGIP from %v to %v or above.", as3SupportedVersion,
-		bigIPAS3Version, as3SupportedVersion)
-}
 
 func (postMgr *PostManager) updateTenantResponseMap(agentWorkerUpdate bool) {
 	/*
@@ -787,101 +404,37 @@ func (postMgr *PostManager) updateRetryMap(tenant string, resp tenantResponse, t
 	}
 }
 
-func (postMgr *PostManager) pollTenantStatus(agentWorkerUpdate bool) {
-
-	var acceptedTenants []string
-	// Create a set to hold unique polling ids
-	acceptedTenantIds := map[string]struct{}{}
-
-	postMgr.tenantResponseMap = make(map[string]tenantResponse)
-
-	for tenant, cfg := range postMgr.retryTenantDeclMap {
-		// So, when we call updateTenantResponseMap, we have to retain failed agentResponseCodes and taskId's correctly
-		postMgr.tenantResponseMap[tenant] = tenantResponse{agentResponseCode: cfg.agentResponseCode, taskId: cfg.taskId}
-		if cfg.taskId != "" {
-			if _, found := acceptedTenantIds[cfg.taskId]; !found {
-				acceptedTenantIds[cfg.taskId] = struct{}{}
-				acceptedTenants = append(acceptedTenants, tenant)
-			}
-		}
-	}
-
-	for len(acceptedTenantIds) > 0 {
-		// Keep retrying until accepted tenant statuses are updated
-		// This prevents agent from unlocking and thus any incoming post requests (config changes) also need to hold on
-		for taskId := range acceptedTenantIds {
-			<-time.After(timeoutMedium)
-			postMgr.getTenantConfigStatus(taskId)
-		}
-		for _, tenant := range acceptedTenants {
-			acceptedTenantIds = map[string]struct{}{}
-			// Even if there is any pending tenant which is not updated, keep retrying for that ID
-			if postMgr.tenantResponseMap[tenant].taskId != "" {
-				acceptedTenantIds[postMgr.tenantResponseMap[tenant].taskId] = struct{}{}
-			}
-		}
-	}
-
-	if len(acceptedTenants) > 0 {
-		postMgr.updateTenantResponseMap(agentWorkerUpdate)
-	}
-}
-
-func (postMgr *PostManager) createAS3Declaration(tenantDeclMap map[string]as3Tenant, userAgent string) as3Declaration {
-	var as3Config map[string]interface{}
-
-	baseAS3ConfigTemplate := fmt.Sprintf(baseAS3Config, postMgr.AS3VersionInfo.as3Version, postMgr.AS3VersionInfo.as3Release, postMgr.AS3VersionInfo.as3SchemaVersion)
-	_ = json.Unmarshal([]byte(baseAS3ConfigTemplate), &as3Config)
-
-	adc := as3Config["declaration"].(map[string]interface{})
-
-	controlObj := make(map[string]interface{})
-	controlObj["class"] = "Controls"
-	controlObj["userAgent"] = userAgent
-	adc["controls"] = controlObj
-
-	for tenant, decl := range tenantDeclMap {
-		adc[tenant] = decl
-	}
-	decl, err := json.Marshal(as3Config)
-	if err != nil {
-		log.Debugf("[AS3] Unified declaration: %v\n", err)
-	}
-
-	return as3Declaration(decl)
-}
-
-func (postMgr *PostManager) retryFailedTenant(userAgent string) {
-	var retryTenants []string
-
-	// this map is to collect all non-201 tenant configs
-	retryDecl := make(map[string]as3Tenant)
-
-	postMgr.tenantResponseMap = make(map[string]tenantResponse)
-
-	for tenant, cfg := range postMgr.retryTenantDeclMap {
-		// So, when we call updateTenantResponseMap, we have to retain failed agentResponseCodes and taskId's correctly
-		postMgr.tenantResponseMap[tenant] = tenantResponse{agentResponseCode: cfg.agentResponseCode, taskId: cfg.taskId}
-		if cfg.taskId == "" {
-			retryTenants = append(retryTenants, tenant)
-			retryDecl[tenant] = cfg.as3Decl.(as3Tenant)
-		}
-	}
-
-	if len(retryTenants) > 0 {
-		// Until all accepted tenants are not processed, we do not want to re-post failed tenants since we will anyways get a 503
-		cfg := agentConfig{
-			data:      string(postMgr.createAS3Declaration(retryDecl, userAgent)),
-			as3APIURL: postMgr.getAS3APIURL(retryTenants),
-			id:        0,
-		}
-
-		postMgr.postConfig(&cfg)
-
-		postMgr.updateTenantResponseMap(false)
-	}
-
-}
+//func (postMgr *PostManager) retryFailedTenant(userAgent string) {
+//	var retryTenants []string
+//
+//	// this map is to collect all non-201 tenant configs
+//	retryDecl := make(map[string]as3Tenant)
+//
+//	postMgr.tenantResponseMap = make(map[string]tenantResponse)
+//
+//	for tenant, cfg := range postMgr.retryTenantDeclMap {
+//		// So, when we call updateTenantResponseMap, we have to retain failed agentResponseCodes and taskId's correctly
+//		postMgr.tenantResponseMap[tenant] = tenantResponse{agentResponseCode: cfg.agentResponseCode, taskId: cfg.taskId}
+//		if cfg.taskId == "" {
+//			retryTenants = append(retryTenants, tenant)
+//			retryDecl[tenant] = cfg.as3Decl.(as3Tenant)
+//		}
+//	}
+//
+//	if len(retryTenants) > 0 {
+//		// Until all accepted tenants are not processed, we do not want to re-post failed tenants since we will anyways get a 503
+//		cfg := agentConfig{
+//			data:      string(postMgr.createAS3Declaration(retryDecl, userAgent)),
+//			as3APIURL: postMgr.getAS3APIURL(retryTenants),
+//			id:        0,
+//		}
+//
+//		postMgr.postConfig(&cfg)
+//
+//		postMgr.updateTenantResponseMap(false)
+//	}
+//
+//}
 
 // function for returning the prefix string for request id
 func getRequestPrefix(id int) string {
