+++ pkg/controller/types.go
@@ -53,12 +53,12 @@ type (
 	Controller struct {
 		mode                        ControllerMode
 		resources                   *ResourceStore
+		RequestHandler              *RequestHandler
 		ciliumTunnelName            string
 		vxlanMgr                    *vxlan.VxlanMgr
 		initialResourceCount        int
 		resourceQueue               workqueue.RateLimitingInterface
 		Partition                   string
-		Agent                       *Agent
 		PoolMemberType              string
 		UseNodeInternal             bool
 		initState                   bool
@@ -85,6 +85,7 @@ type (
 		discoveryMode               discoveryMode
 		clusterRatio                map[string]*int
 		clusterAdminState           map[string]clustermanager.AdminState
+		APIHandler                  APIHandlerInterface
 		resourceContext
 	}
 	resourceContext struct {
@@ -161,9 +162,9 @@ type (
 	Params struct {
 		Config                      *rest.Config
 		Namespaces                  []string
+		RequestHandler              *RequestHandler
 		NamespaceLabel              string
 		Partition                   string
-		Agent                       *Agent
 		PoolMemberType              string
 		VXLANName                   string
 		VXLANMode                   string
@@ -807,21 +808,108 @@ type (
 		*PostManager
 		Partition string
 	}
+
+	RequestHandler struct {
+		AgentWorkers []*AgentWorker
+		reqChan      chan ResourceConfigRequest
+		userAgent    string
+		respChan     chan resourceStatusMeta
+		//CMTokenManager                  *tokenmanager.TokenManager
+		ccclGTMAgent bool
+		disableARP   bool
+		HAMode       bool
+	}
+
+	AgentWorker struct {
+		*Agent
+		Type              string
+		httpClientMetrics bool
+		stopChan          chan struct{}
+		BigIpAddress      string
+		PythonDriverPID   int
+		postChan          chan ResourceConfigRequest
+		StopChan          chan interface{}
+	}
+
 	Agent struct {
+		*APIHandler
+		Partition                       string
+		PrimaryClusterHealthProbeParams PrimaryClusterHealthProbeParams
+		ConfigWriter                    writer.Writer
+		EventChan                       chan interface{}
+		respChan                        chan resourceStatusMeta
+		PythonDriverPID                 int
+		userAgent                       string
+		HttpAddress                     string
+		EnableIPV6                      bool
+		declUpdate                      sync.Mutex
+		ccclGTMAgent                    bool
+		disableARP                      bool
+		HAMode                          bool
+	}
+
+	BaseAPIHandler struct {
+		apiType    string
+		APIHandler APIHandlerInterface
+		Partition  string
 		*PostManager
-		Partition       string
-		ConfigWriter    writer.Writer
-		EventChan       chan interface{}
-		respChan        chan resourceStatusMeta
-		PythonDriverPID int
-		userAgent       string
-		HttpAddress     string
-		EnableIPV6      bool
-		declUpdate      sync.Mutex
-		ccclGTMAgent    bool
-		disableARP      bool
-		HAMode          bool
-		GTMPostManager  *GTMPostManager
+	}
+
+	GTMAPIHandler struct {
+		*BaseAPIHandler
+		Partition string
+	}
+
+	LTMAPIHandler struct {
+		*BaseAPIHandler
+	}
+
+	APIHandler struct {
+		GTM *GTMAPIHandler
+		LTM *LTMAPIHandler
+	}
+
+	// PostManager functionality. Embedding PostManager in AS3Handler would limit reusability across
+	// other API types like GTM. The current hierarchy allows:
+	// 1. Common HTTP posting capabilities via PostManager
+	// 2. API-specific handling via apiHandler interface
+	// 3. Specialized GTM posting via GTMPostManager
+	// This separation of concerns is appropriate for the controller architecture.
+
+	AS3Handler struct {
+		AS3Config         map[string]interface{}
+		AS3VersionInfo    as3VersionInfo
+		bigIPAS3Version   float64
+		postManagerPrefix string
+		*PostParams
+		*AS3Parser
+	}
+
+	AS3Parser struct {
+		// AS3Parser implements AS3ParserInterface
+		AS3ParserInterface
+		AS3VersionInfo  as3VersionInfo
+		bigIPAS3Version float64
+	}
+
+	PostManager struct {
+		httpClient        *http.Client
+		tenantResponseMap map[string]tenantResponse
+		PostParams
+		PrimaryClusterHealthProbeParams PrimaryClusterHealthProbeParams
+		firstPost                       bool
+		postManagerPrefix               string
+		// cachedTenantDeclMap,incomingTenantDeclMap hold tenant names and corresponding AS3 config
+		cachedTenantDeclMap   map[string]as3Tenant
+		incomingTenantDeclMap map[string]as3Tenant
+		// this map stores the tenant priority map
+		tenantPriorityMap map[string]int
+		// retryTenantDeclMap holds tenant name and its agent Config,tenant details
+		retryTenantDeclMap map[string]*tenantParams
+		postChan           chan ResourceConfigRequest
+		respChan           chan resourceStatusMeta
+		httpClientMetrics  bool
+		retryChan          chan struct{}
 	}
 
 	AgentParams struct {
@@ -842,28 +930,11 @@ type (
 		StaticRoutingMode  bool
 		SharedStaticRoutes bool
 		MultiClusterMode   string
+		ApiType            string
+		PrimaryBigIP       string
+		SecondaryBigIP     string
+		HAMode             bool
 	}
-
-	PostManager struct {
-		httpClient        *http.Client
-		tenantResponseMap map[string]tenantResponse
-		PostParams
-		PrimaryClusterHealthProbeParams PrimaryClusterHealthProbeParams
-		firstPost                       bool
-		AS3VersionInfo                  as3VersionInfo
-		bigIPAS3Version                 float64
-		postManagerPrefix               string
-		// cachedTenantDeclMap,incomingTenantDeclMap hold tenant names and corresponding AS3 config
-		cachedTenantDeclMap   map[string]as3Tenant
-		incomingTenantDeclMap map[string]as3Tenant
-		// this map stores the tenant priority map
-		tenantPriorityMap map[string]int
-		// retryTenantDeclMap holds tenant name and its agent Config,tenant details
-		retryTenantDeclMap map[string]*tenantParams
-		postChan           chan ResourceConfigRequest
-		retryChan          chan struct{}
-	}
-
 	PrimaryClusterHealthProbeParams struct {
 		paramLock     *sync.RWMutex
 		EndPoint      string
@@ -882,8 +953,8 @@ type (
 		SSLInsecure   bool
 		AS3PostDelay  int
 		// Log the AS3 response body in Controller logs
-		LogAS3Response    bool
-		LogAS3Request     bool
+		LogResponse       bool
+		LogRequest        bool
 		HTTPClientMetrics bool
 	}
 
@@ -939,7 +1010,7 @@ type (
 	// AS3 version struct
 
 	as3VersionInfo struct {
-		as3Version       string
+		as3Version       float64
 		as3SchemaVersion string
 		as3Release       string
 	}
